populate the board with todos

+++ src/boards/boards.controller.ts
@@ -3,34 +3,54 @@ import {
   Controller,
   Delete,
   Get,
+  HttpException,
   Param,
+  Patch,
   Post,
-  Put,
+  Query,
 } from '@nestjs/common';
 import { BoardsService } from './boards.service';
 import UpdateTodoDto from './dto/update-board.dto';
+import mongoose from 'mongoose';
+import { CreateBoardDto } from './dto/create-board.dto';
 
 @Controller('boards')
 export class BoardsController {
   constructor(private readonly boardsService: BoardsService) {}
 
   @Post()
-  addBoard(@Body('name') boardName: string) {
-    return this.boardsService.create(boardName);
+  addBoard(@Body() board: CreateBoardDto) {
+    return this.boardsService.create(board);
   }
 
   @Get()
-  getAllBoards() {
-    return this.boardsService.findAll();
+  getAllBoards(@Query('id') boardId?: string) {
+    if (boardId) {
+      const isValid = mongoose.Types.ObjectId.isValid(boardId);
+      if (!isValid) {
+        throw new HttpException('Invalid ID', 400);
+      }
+      return this.boardsService.findOne(boardId);
+    } else {
+      return this.boardsService.findAll();
+    }
   }
 
   @Delete(':id')
   deleteBoardById(@Param('id') boardId: string) {
+    const isValid = mongoose.Types.ObjectId.isValid(boardId);
+    if (!isValid) {
+      throw new HttpException('Invalid ID', 400);
+    }
     return this.boardsService.deleteById(boardId);
   }
 
-  @Put(':id')
+  @Patch(':id')
   updateBoard(@Param('id') boardId: string, @Body() board: UpdateTodoDto) {
+    const isValid = mongoose.Types.ObjectId.isValid(boardId);
+    if (!isValid) {
+      throw new HttpException('Invalid ID', 400);
+    }
     return this.boardsService.updateById(boardId, board);
   }
 }
